Alterações na página do feed e post
